Fix build failures from a clean checkout
The compilation recipe expects the presence of a `bin` directory to
emit the output binary. That directory does not exist in a clean
checkout of the repository, causing the link recipe to fail. Fix that
recipe to make that directory if it doesn't already exist.

Also ignore the build products in the `bin` directory in `.gitignore`.

The `Makefile` allows the `test` target rule to run parallel to the
`$(TARGETS)` rule (especially when recipe parallelization is enabled),
which results in several test failures due to the programs not yet
linked. Fix this by enlisting `$(TARGETS)` as a prerequisite of `test`.

Finally, fix a minor warning in `mcp19` due to GCC thinking a variable
gets used without initialization. While GCC is incorrect here, the
warning does cause the build to fail.